OpticalFibers - Common structs and functions

Convenient functions

OpticalFibers.piecewiseIndex — Function

piecewiseIndex(pos::Real,r::Union{AbstractVector{<:Real},Real},n::AbstractVector{<:Real})

Function that returns the refractive index at the position pos in the case of a step-index fiber profile defined by the vectors r and n.

OpticalFibers.meshgrid — Function

meshgrid(v::AbstractVector)

Function equivalent to the matlab function

meshgrid(vx::AbstractVector{T}, vy::AbstractVector{T})

Function equivalent to the matlab function

OpticalFibers.derivative — Function

derivative(t::Tuple{Vector{<:Real},Vector{<:Number}})

Function that returns the derivative of the tuple t=(x,y)

derivative(t::Tuple{Vector{<:Real},Vector{<:Number}},order::Int64)

Function that returns the nth order derivative of the tuple t=(x,y)

OpticalFibers.ring — Function

ring(N::Integer)

Function that returns a tuple of vectors containing the coordinates (x,y) of the holes in the Nth ring of a PCF

Tensor

OpticalFibers.tensor3 — Type

tensor

Structure that describes a tensor of permittivity or permeability. Each of the 9 components of the a tensor3 is a function of a tuple (x,y).

xx :: Function

yx :: Function

zx :: Function

xy :: Function

yy :: Function

zy :: Function

xz :: Function

yz :: Function

zz :: Function

OpticalFibers.tensor3 — Method

tensor3(xx::Union{Function,Number},yx::Union{Function,Number},zx::Union{Function,Number},xy::Union{Function,Number},yy::Union{Function,Number},zy::Union{Function,Number},xz::Union{Function,Number},yz::Union{Function,Number},zz::Union{Function,Number})

Function that returns a tensor3 defined with functions and constant values. Each function must always return the same type.

OpticalFibers.tensor3 — Method

tensor3(fxx::Union{Function,Number},fyy::Union{Function,Number},fzz::Union{Function,Number})

Function that returns a diagonal tensor3

OpticalFibers.tensor3 — Method

tensor3(f::Union{Function,Number})

Function that returns a diagonal tensor3 with the same function in the three diagonal terms

OpticalFibers.inverse — Function

inverse(t::tensor3)

Returns the inverse of the tensor t