OpticalFibers

Constants

OpticalFibers.PhysicalData.Z0 — Constant

Z0=sqrt(mu0/eps0)

Vaccum impedance (in Ω)

OpticalFibers.PhysicalData.Z0_Unitful — Constant

Z0_Unitful=sqrt(mu0_Unitful/eps0_Unitful)

Vaccum impedance

OpticalFibers.PhysicalData.c — Constant

c=299792458.0

Velocity of light in vacuum (in m/s)

OpticalFibers.PhysicalData.c_Unitful — Constant

c_Unitful=299792458.0*u"m/s"

Velocity of light in vacuum

OpticalFibers.PhysicalData.eps0 — Constant

eps0=1/(mu0*c^2)

Vacuum permittivity (in F/m)

OpticalFibers.PhysicalData.eps0_Unitful — Constant

eps0_Unitful=1/(mu0_Unitful*c_Unitful^2)

Vacuum permittivity

OpticalFibers.PhysicalData.h — Constant

h=6.62607015e-34

Planck constant (in J.s)

OpticalFibers.PhysicalData.h_Unitful — Constant

h_Unitful=6.62607015e-34*u"J*s"

Planck constant

OpticalFibers.PhysicalData.mu0 — Constant

mu0=1.25663706212e-6

Vacuum permeability (in H/m)

OpticalFibers.PhysicalData.mu0_Unitful — Constant

mu0_Unitful=1.25663706212e-6*u"H/m"

Vacuum permeability

Functions

OpticalFibers.PhysicalData.Sa_Ytterbium — Method

Sa_Ytterbium(lambda::Union{Real,Quantity{<:Real,Unitful.𝐋}})

Returns the absorption cross section (in m²) of ytterbium at wavelength lambda (in meters if lambda is a Real) [2]

OpticalFibers.PhysicalData.n_F_Doped_Silica — Method

n_F_Doped_Silica(lambda::Union{Real,Quantity{<:Real,Unitful.𝐋}},xF::Real)

Returns the refractive index of fluorine-doped silica with a fluorine fraction xF (0 ≤ xGe ≤ 1) at the wavelength lambda (in meters if lambda is a Real) [3].

Domain of validity: 600 nm ≤lambda≤ 1800 nm and xF ≤ 0.2

OpticalFibers.PhysicalData.n_Ge_Doped_Silica — Method

n_Ge_Doped_Silica(lambda::Union{Real,Quantity{<:Real,Unitful.𝐋}},xGe::Real;author::Symbol=:Fleming)

Returns the refractive index of germanium-doped silica with a germanium fraction xGe (0 ≤ xGe ≤ 1) at the wavelength lambda (in meters if lambda is a Real). Two Sellmeier formulas can be used : author=:Fleming [4] and author=:Sunak [3].

Domain of validity:

360 nm ≤lambda≤ 4300 nm for Fleming's formula

600 nm ≤lambda≤ 1800 nm and xGe ≤ 0.2 for Sunak's formula

OpticalFibers.PhysicalData.n_Germanium — Method

n_Germanium(lambda::Real,T::Real)

Returns the refractive index of Germanium at wavelength lambda (in meters if lambda is a Real) and at temperature T (in Kelvin if T is a Real) [5]

Domain of validity: 1900 nm ≤ lambda ≤ 5500 nm and 20 K ≤ T ≤ 300 K

OpticalFibers.PhysicalData.n_Silicon — Method

n_Silicon(lambda::Union{Real,Quantity{<:Real,Unitful.𝐋}},T::Union{Real,Quantity{<:Real,Unitful.𝚯}})

Returns the refractive index of Silicon at wavelength lambda (in meters if lambda is a Real) and at temperature T (in Kelvin if T is a Real) [5]

Domain of validity: 1100 nm ≤ lambda ≤ 5600 nm and 20 K ≤ T ≤ 300 K