OpticalFibers - Common structs and functions

Convenient functions

OpticalFibers.piecewiseIndex — Function

piecewiseIndex(pos::Real,r::Union{AbstractVector{<:Real},Real},n::AbstractVector{<:Real})

Function that returns the refractive index at the position pos in the case of a step-index fiber profile defined by the vectors r and n.

OpticalFibers.meshgrid — Function

meshgrid(v::AbstractVector)

Function equivalent to the matlab function

meshgrid(vx::AbstractVector{T}, vy::AbstractVector{T})

Function equivalent to the matlab function

OpticalFibers.derivative — Function

derivative(t::Tuple{Vector{<:Real},Vector{<:Number}})

Function that returns the derivative dy/dx of the tuple t=(x,y)

derivative(t::Tuple{Vector{<:Real},Vector{<:Number}},order::Int64)

Function that returns the nth order derivative dy/dx of the tuple t=(x,y)

OpticalFibers.ring — Function

ring(N::Integer)

Function that returns a tuple of vectors containing the coordinates (x,y) of the holes in the Nth ring of a PCF

OpticalFibers.approx_nFSM_PCF — Function

approx_nFSM_PCF(lambda::Real,D::Real,pitch::Real)

Returns an approximate value of nFSM for photonic crystal fiber [1]

lambda: Wavelength

D: Hole diameter

pitch: Pitch

OpticalFibers.approx_neff_PCF — Function

approx_neff_PCF(lambda::Real,D::Real,pitch::Real)

Returns an approximate value of the effective index of the fundamental mode of a photonic crystal fiber [1]

lambda: Wavelength

D: Hole diameter

pitch: Pitch

OpticalFibers.add_cylindrical_PML — Function

add_cylindrical_PML(epsmu::Union{Number,Function,tensor3},r_pml::Real,d_pml::Real,alpha::Real)

Function that returns a tensor of permittivity/permeability with a cylindrical PML

epsmu: permittivity/permeability profile of the fiber. The fiber is assumed to be isotropic when epsmu is a Number or a Function

r_pml: distance between the fiber center and the PML beginning

d_pml: PML thickness

alphaPML: attenuation factor of the PML

OpticalFibers.add_rectangular_PML — Function

add_rectangular_PML(epsmu::Union{Number,Function,tensor3},x_pml::Real,dx_pml::Real,y_pml::Real,dy_pml::Real,alpha::Real)

Function that returns a tensor of permittivity/permeability with a rectangular PML located at [xpml,xpml+dxpml], [-xpml-dxpml,-xpml], [ypml,ypml+dypml], [-ypml-dypml,-ypml]

epsmu: permittivity/permeability profile of the fiber. The fiber is assumed to be isotropic when epsmu is a Number or a Function

x_pml: distance between the fiber center and the PML beginning in the x direction

dx_pml: PML thickness in the x direction

y_pml: distance between the fiber center and the PML beginning in the y direction

dy_pml: PML thickness in the y direction

alphaPML: attenuation factor of the PML

add_rectangular_PML(epsmu::Union{Number,Function,tensor3},xm_pml::Real,xp_pml::Real,dx_pml::Real,ym_pml::Real,yp_pml::Real,dy_pml::Real,alphaPML::Real)

Function that returns a tensor of permittivity/permeability with a rectangular PML located at [xppml,xppml+dxpml], [-xmpml-dxpml,-xmpml], [yppml,yppml+dypml], [-ympml-dypml,-ympml]

epsmu: permittivity/permeability profile of the fiber

xm_pml: minimum position in the x-direction of the beginning of the PML

xp_pml: maximum position in the x-direction of the beginning of the PML

dx_pml: PML thickness in the x direction

ym_pml: minimum position in the y-direction of the beginning of the PML

yp_pml: maximum position in the y-direction of the beginning of the PML

dy_pml: PML thickness in the y direction

alphaPML: attenuation factor of the PML

OpticalFibers.add_twist_PML — Function

add_twist_PML(epsmu::Union{Function,Number},P::Real,r_pml::Real,d_pml::Real,alphaPML::Real)

Function that returns a tensor of permittivity/permeability of a twisted isotropic fiber with a cylindrical PML

epsmu: function of the tuple (x,y) that describes the permittivity/permeability profile of the fiber

P: period of the twist

r_pml: distance between the fiber center and the PML beginning

d_pml: PML thickness

alpha: attenuation factor of the PML

OpticalFibers.integrate1D — Function

integrate1D(f::Function;rtol::Real=1E-5;kwargs...)

Integrate a 1D function f(r) from 0 to infinity

The keyword arguments are those of HCubature.jl. In particular,rtol and atol are the relative and the absolute tolerance on the result

OpticalFibers.integrate2D — Function

integrate2D(f::Function;kwargs...)

Integrate a 2D function f(r) with r is a the tuple (x,y) for x and y from -infinity to infinity

The keyword arguments are those of HCubature.jl. In particular,`rtol` and `atol` are the relative and the absolute tolerance on the result

OpticalFibers.compute_kt — Function

compute_kt(KNumber::Int,CellType::Symbol;Irreducible::Bool=true,Pitch::Real=1,MeshType::Symbol=:Internal)

Compute the vectors of a 2D Brillouin zone

KNumber: Integer related to the number of vectors to compute

CellType: :Square or :Hexagon

Irreducible: true-> irreducible Brillouin zone, false->entire Billouin zone

pitch: Pitch

MeshType: Internal (useful to compute the Density Of States) or :Edge (useful to compute the bandgap edge)

Tensor

OpticalFibers.tensorComponent — Type

tensorComponent

Structure that describes a the component of the tensor of permittivity or permeability.

f::Union{Function,Number}

OpticalFibers.tensor3 — Type

tensor

Structure that describes a tensor of permittivity or permeability. Each of the 9 components of the a tensor3 is a function of a tuple (x,y).

xx :: tensorComponent

yx :: tensorComponent

zx :: tensorComponent

xy :: tensorComponent

yy :: tensorComponent

zy :: tensorComponent

xz :: tensorComponent

yz :: tensorComponent

zz :: tensorComponent

OpticalFibers.inverse — Function

inverse(t::tensor3)

Returns the inverse of the tensor t

LinearAlgebra.det — Function

det(t::tensor3)

Returns the determinant of the tensor t