OpticalFibers - Common structs and functions

Mathematical functions

OpticalFibers.meshgrid — Function

meshgrid(v::AbstractVector)

Function equivalent to the matlab function

meshgrid(vx::AbstractVector{T}, vy::AbstractVector{T})

Function equivalent to the matlab function

OpticalFibers.function_integrate — Function

function_integrate(f::Function,a::AbstractVector,b::AbstractVector;kwargs...)

Returns the integral $\int_{a_1}^{b_1} \int_{a_2}^{b_2} ... f(x) dx$ where x is a unitless Point. The integration is performed with a change of variable x=tan(t) and the hcubature function of the HCubature.jl package.

a: Vector of the lower bounds of integration

b: Vector of the upper bounds of integration

kwargs: keyword arguments of the hcubature function (atol, rtol, initdiv, norm)

OpticalFibers.function_integrate_unitful — Function

function_integrate_unitful(f::Function,a::AbstractVector,b::AbstractVector;characteristic_length::AbstractVector=[],kwargs...)

Returns the integral $\int_{a_1}^{b_1} \int_{a_2}^{b_2} ... f(x) dx$ where x is a Point of dimension length(a) with unit handling. The integration is performed with a change of variable x=tan(t) and the hcubature function of the HCubature.jl package.

a: Vector of the lower bounds of integration

b: Vector of the upper bounds of integration

characteristic_length: Vector of characteristic lengths to make the integration bounds unitless and significant variations around 1. If not provided, the unit of a is used.

kwargs: keyword arguments of the hcubature function (atol, rtol, initdiv, norm)

Functions for optical fibers

OpticalFibers.piecewiseIndex — Function

piecewiseIndex(pos::T,r::Union{AbstractVector{<:T2},T2},n::AbstractVector) where {T<:Union{Real,realQuantity},T2<:Union{Real,realQuantity}}

Function that returns the refractive index at the position pos in the case of a step-index fiber profile defined by the vectors r and n.

OpticalFibers.ring — Function

ring(N::Integer)

Function that returns a tuple of vectors containing the coordinates (x,y) of the holes in the Nth ring of a PCF (you have to multiply by the pitch to get the real coordinates)

OpticalFibers.approx_nFSM_PCF — Function

approx_nFSM_PCF(lambda::Real,D::Real,pitch::Real)

Returns an approximate value of nFSM for photonic crystal fiber [1]

lambda: Wavelength

D: Hole diameter

pitch: Pitch

OpticalFibers.approx_neff_PCF — Function

approx_neff_PCF(lambda::Real,D::Real,pitch::Real)

Returns an approximate value of the effective index of the fundamental mode of a photonic crystal fiber [1]

lambda: Wavelength

D: Hole diameter

pitch: Pitch

OpticalFibers.add_cylindrical_PML — Function

add_cylindrical_PML(epsmu::Union{Number,Function},r_pml::Union{Real,realLength},d_pml::Union{Real,realLength},alphaPML::Real)

Function that returns a tensor of permittivity/permeability with a cylindrical PML

epsmu: permittivity/permeability profile of the fiber. The fiber is assumed to be isotropic when epsmu is a Number or a Function

r_pml: distance between the fiber center and the PML beginning

d_pml: PML thickness

alphaPML: attenuation factor of the PML

OpticalFibers.add_rectangular_PML — Function

add_rectangular_PML(epsmu::Union{Number,Function},x_pml::Union{Real,realLength},dx_pml::Union{Real,realLength},y_pml::Union{Real,realLength},dy_pml::Union{Real,realLength},alphaPML::Real)

Function that returns a tensor of permittivity/permeability with a rectangular PML located at [xpml,xpml+dxpml], [-xpml-dxpml,-xpml], [ypml,ypml+dypml], [-ypml-dypml,-ypml]

epsmu: permittivity/permeability profile of the fiber. The fiber is assumed to be isotropic when epsmu is a Number or a Function

x_pml: distance between the fiber center and the PML beginning in the x direction

dx_pml: PML thickness in the x direction

y_pml: distance between the fiber center and the PML beginning in the y direction

dy_pml: PML thickness in the y direction

alphaPML: attenuation factor of the PML

add_rectangular_PML(epsmu::Union{Number,Function},xm_pml::Union{Real,realLength},xp_pml::Union{Real,realLength},dx_pml::Union{Real,realLength},ym_pml::Union{Real,realLength},yp_pml::Union{Real,realLength},dy_pml::Union{Real,realLength},alphaPML::Real)

Function that returns a tensor of permittivity/permeability with a rectangular PML located at [xppml,xppml+dxpml], [-xmpml-dxpml,-xmpml], [yppml,yppml+dypml], [-ympml-dypml,-ympml]

epsmu: permittivity/permeability profile of the fiber

xm_pml: minimum position in the x-direction of the beginning of the PML

xp_pml: maximum position in the x-direction of the beginning of the PML

dx_pml: PML thickness in the x direction

ym_pml: minimum position in the y-direction of the beginning of the PML

yp_pml: maximum position in the y-direction of the beginning of the PML

dy_pml: PML thickness in the y direction

alphaPML: attenuation factor of the PML

OpticalFibers.add_twist_PML — Function

add_twist_PML(epsmu::Union{Number,Function},P::Union{Real,realLength},r_pml::Union{Real,realLength},d_pml::Union{Real,realLength},alphaPML::Real)

Function that returns a tensor of permittivity/permeability of a twisted isotropic fiber with a cylindrical PML

epsmu: function of the tuple (x,y) that describes the permittivity/permeability profile of the fiber

P: period of the twist

r_pml: distance between the fiber center and the PML beginning

d_pml: PML thickness

alpha: attenuation factor of the PML

OpticalFibers.compute_kt — Function

compute_kt(KNumber::Int,CellType::Symbol;Irreducible::Bool=true,Pitch::Real=1,MeshType::Symbol=:Internal)

Compute the vectors of a 2D Brillouin zone

KNumber: Integer related to the number of vectors to compute

CellType: :Square or :Hexagon

Irreducible: true-> irreducible Brillouin zone, false->entire Billouin zone

pitch: Pitch

MeshType: Internal (useful to compute the Density Of States) or :Edge (useful to compute the bandgap edge)