OpticalFibers.ModeSolvers

This module contains different mode solvers for optical fibers. They are based on the interface conditions of analytical solutions between the layers of the fiber, finite difference method or finite element method.

Fields

This module defines two types of field: the scalar field (valid in the case of weakly guiding fibers) and the vector field.

OpticalFibers.ModeSolvers.Field — Type

Abstract structure to describe an electromagnetic field

abstract type Field end

OpticalFibers.ModeSolvers.ScalarField — Type

Structure describing a 2D scalar field in cartesian coordinates

x :: Vector{Float64} - Horizontal coordinate

y :: Vector{Float64} - Vertical coordinate

E :: Matrix{ComplexF64} - Electric field

OpticalFibers.ModeSolvers.VectorField — Type

Structure describing a 2D vector field in cartesian coordinates

x :: Vector{Float64} - Horizontal coordinate

y :: Vector{Float64} - Vertical coordinate

Ex :: Matrix{ComplexF64} - x-component of the electric field

Ey :: Matrix{ComplexF64} - y-component of the electric field

Ez :: Matrix{ComplexF64} - z-component of the electric field

Hx :: Matrix{ComplexF64} - x-component of the magnetic field

Hy :: Matrix{ComplexF64} - y-component of the magnetic field

Hz :: Matrix{ComplexF64} - z-component of the magnetic field

OpticalFibers.ModeSolvers.ScalarFieldFEM — Type

Structure describing a scalar field defined with a Triangulation

Ω :: Gridap.Triangulation

dΩ :: Gridap.CellData.Measure

E :: Gridap.CellField - Electric field

OpticalFibers.ModeSolvers.VectorFieldFEM — Type

Structure describing a vector field defined with a Triangulation

Ω :: Gridap.Triangulation

dΩ :: Gridap.CellData.Measure

Ex :: Gridap.CellField - x-component of the electric field

Ey :: Gridap.CellField - y-component of the electric field

Ez :: Gridap.CellField - z-component of the electric field

Hx :: Gridap.CellField - x-component of the magnetic field

Hy :: Gridap.CellField - y-component of the magnetic field

Hz :: Gridap.CellField - z-component of the magnetic field

Two fields of the same type can be add or substract if their coordinates x and y are similar. It is also possible to multiply or divide a field by a scalar.

Two convenient functions that interpolate a field on another grid (changing x and y) are implemented:

OpticalFibers.ModeSolvers.Interpolation — Function

Interpolation(f::ScalarField,x::Vector{<:Real},y::Vector{<:Real})

Returns a scalar field obtained by the linear interpolation of the field f.

This function assumes that the electric field is null outside the box defined by f.x and f.y

Interpolation(f::VectorField,x::Vector{<:Real},y::Vector{<:Real})

Returns a vector field obtained by the linear interpolation of the field f.

This function assumes that the electric/magnetic field is null outside the box defined by f.x and f.y

Modes

OpticalFibers.ModeSolvers.Mode — Type

Abstract structure to describe an optical fiber mode

abstract type Mode end

OpticalFibers.ModeSolvers.ScalarMode1D — Type

Structure describing a scalar mode of an optical fiber with a revolution symmetry

Name :: String - Name of the mode

neff :: Float64 - Effective index

lambda :: Float64 - the wavelength at which the mode was calculated

nu :: Int64 - Azimuthal number

r :: Vector{Float64} - Radial coordinate

E :: Vector{Float64} - Electric field

OpticalFibers.ModeSolvers.ScalarMode2D — Type

Structure describing a scalar mode of an optical fiber in cartesian coordinates

Name :: String - Name of the mode

neff :: Float64 - Effective index

lambda :: Float64 - the wavelength at which the mode was calculated

x :: Vector{Float64} - Horizontal coordinate

y :: Vector{Float64} - Vertical coordinate

E :: Matrix{Float64} - Electric field

OpticalFibers.ModeSolvers.VectorMode — Type

Structure describing a vector mode of an optical fiber in cartesian coordinates.

This structure assumes that the fiber is made of a non-lossy materials. In this case, the x- and y-components of the electric/magnetic fields can be chosen real and their z-components are then pure imaginary numbers.

Name :: String - Name of the mode

neff :: Float64 - Effective index

lambda :: Float64 - the wavelength at which the mode was calculated

x :: Vector{Float64} - Horizontal coordinate

y :: Vector{Float64} - Vertical coordinate

Ex :: Matrix{Float64} - Real part of the x-component of the electric field

Ey :: Matrix{Float64} - Real part of the y-component of the electric field

Ez :: Matrix{Float64} - Imaginary part of the z-component of the electric field

Hx :: Matrix{Float64} - Real part of the x-component of the magnetic field

Hy :: Matrix{Float64} - Real part of the y-component of the magnetic field

Hz :: Matrix{Float64} - Imaginary part of the z-component of the magnetic field

OpticalFibers.ModeSolvers.ScalarModeFEM — Type

Structure describing a scalar mode of an optical fiber computed with FEM.

Name :: String - Name of the mode

neff :: Float64 - Effective index

lambda :: Float64 - the wavelength at which the mode was calculated

Ω :: Gridap.Triangulation

dΩ :: Gridap.CellData.Measure

E :: Gridap.CellField - Electric field

OpticalFibers.ModeSolvers.VectorModeFEM — Type

Structure describing a vector mode of an optical fiber computed with FEM.

Name :: String - Name of the mode

neff :: Float64 - Effective index

lambda :: Float64 - the wavelength at which the mode was calculated

Ω :: Gridap.Triangulation

dΩ :: Gridap.CellData.Measure

Ex :: Gridap.CellField - x-component of the electric field

Ey :: Gridap.CellField - y-component of the electric field

Ez :: Gridap.CellField - z-component of the electric field

Hx :: Gridap.CellField - x-component of the magnetic field

Hy :: Gridap.CellField - y-component of the magnetic field

Hz :: Gridap.CellField - z-component of the magnetic field

A ScalarMode1D can be convert into a ScalarMode2D or a VectorMode:

OpticalFibers.ModeSolvers.ScalarMode2D — Method

ScalarMode2D(m::ScalarMode1D;sincos::Char='c')

Convert a ScalarMode1D into a ScalarMode2D

sincos must be 'c' for a field in cos(nu.θ) or 's' for a field in sin(nu.θ) if m.nu≠0

OpticalFibers.ModeSolvers.VectorMode — Method

VectorMode(m::ScalarMode1D;polar::Char='x',sincos::Char='c')

Convert a ScalarMode1D into a x- or y-polarized VectorMode

polar must be 'x' or 'y'

sincos must be 'c' for a field in cos(nu.θ) or 's' for a field in sin(nu.θ) if m.nu≠0

A ScalarMode2D can be convert into a VectorMode:

OpticalFibers.ModeSolvers.VectorMode — Method

VectorMode(m::ScalarMode2D;polar::Char='x')

Convert a ScalarMode2D into a x- or y-polarized VectorMode

polar must be 'x' or 'y'

A ScalarModeFEM can be convert into a VectorModeFEM:

OpticalFibers.ModeSolvers.VectorModeFEM — Method

VectorModeFEM(m::ScalarModeFEM;polar::Char='x')

Convert a ScalarModeFEM into a x- or y-polarized VectorModeFEM

polar must be 'x' or 'y'

Losses

The mode losses are computed in dB/km from the imaginary part of the effective index.

OpticalFibers.ModeSolvers.losses — Method

losses(m::Mode)

Return the losses of a mode in dB/km if the wavelength is in meters. If the wavelength is in microns, you have to multiply the result by 1E6 to obtain the losses in dB/km.

Propagation of modes

The field due to the propagation of a mode after a propagation distance z can be compute be using these functions:

OpticalFibers.ModeSolvers.ScalarField — Method

ScalarField(m::ScalarMode1D,z::Real=0;sincos::Char='c')

Returns the scalar field due to the mode m after a propagation distance z

sincos must be 'c' for a field in cos(nu.θ) or 's' for a field in sin(nu.θ) if m.nu≠0

z: Propagation distance - Must be in the same unit as m.lambda

OpticalFibers.ModeSolvers.ScalarField — Method

ScalarField(m::ScalarMode2D,z::Real=0)

Returns the scalar field due to the mode m after a propagation distance z

z: Propagation distance - Must be in the same unit as m.lambda

OpticalFibers.ModeSolvers.VectorField — Method

VectorField(m::VectorMode,z::Real=0)

Returns the vector field due to the mode m after a propagation distance z

z: Propagation distance - Must be in the same unit as m.lambda

OpticalFibers.ModeSolvers.VectorField — Method

VectorField(m::ScalarMode2D,z::Real=0;polar::Char='x')

Returns the vector field due to the mode m after a propagation distance z

polar must be 'x' or 'y'

z: Propagation distance - Must be in the same unit as m.lambda

OpticalFibers.ModeSolvers.VectorField — Method

VectorField(m::ScalarMode1D,z::Real=0;polar::Char='x',sincos::Char='c')

Returns the vector field due to the mode m after a propagation distance z

polar must be 'x' or 'y'

sincos must be 'c' for a field in cos(nu.θ) or 's' for a field in sin(nu.θ) if m.nu≠0

z: Propagation distance - Must be in the same unit as m.lambda

OpticalFibers.ModeSolvers.ScalarFieldFEM — Method

ScalarFieldFEM(m::ScalarModeFEM,z::Real=0)

Returns the scalar field due to the mode m after a propagation distance z

z: Propagation distance - Must be in the same unit as m.lambda

OpticalFibers.ModeSolvers.VectorFieldFEM — Method

VectorFieldFEM(m::VectorModeFEM,z::Real=0)

Returns the vector field due to the mode m after a propagation distance z

z: Propagation distance - Must be in the same unit as m.lambda

Poynting vector

Convenient functions that compute the Poynting vector of modes and fields. All functions return a tuple (Px,Py,Pz) that contains the three components of the Poynting vector.

OpticalFibers.ModeSolvers.PoyntingVector — Function

PoyntingVector(f::VectorField)

Return a tuple of 3 matrices.

PoyntingVector(m::VectorMode)

Return a tuple of 3 matrices.

PoyntingVector(m::ScalarMode1D;sincos::Char='c')

Return a tuple of 3 matrices.

sincos must be 'c' for a field in cos(nu.θ) or 's' for a field in sin(nu.θ) if m.nu≠0

PoyntingVector(m::ScalarMode2D)

Return a tuple of 3 matrices.

PoyntingVector(f::VectorModeFEM)

Return a tuple of 3 CellFields.

PoyntingVector(f::ScalarModeFEM)

Return a tuple of 3 CellFields.

Mode normalization

The modes given by the mode solvers of this package are not normalized. You can normalize them by using the function normalize!. The vector modes are normalized so as $\vert \iint \frac{1}{2}\left(\vec{E}\wedge\vec{H}^*\right) dS \vert=1$ and the scalar mode are normalized like vector modes if the optional parameter unitIntegral is false or so as $ \iint \vert E \vert^2 dS=1 $ if unitIntegral is true.

OpticalFibers.ModeSolvers.normalize! — Function

normalize!(m::ScalarMode1D;unitIntegral::Bool=true)

normalize!(m::ScalarMode2D;unitIntegral::Bool=true)

normalize!(m::VectorMode)

normalize!(m::ScalarModeFEM;unitIntegral::Bool=true)

normalize!(m::VectorModeFEM)

Overlap integral

The overlap integral between two fields f$_1$ and f$_2$ is $\langle f_1 \vert f_2 \rangle = \iint E_1 E_2^* dS$ for scalar fields and $\langle f_1 \vert f_2 \rangle = \iint \frac{1}{2}\left(\vec{E}_1 \wedge \vec{H}_2^*\right)_z dS$ for vector fields.

It is also possible to compute overlap integrals between two modes or between a mode and a field.

When modes are involved, they are normalized before the calculation.

OpticalFibers.ModeSolvers.overlap — Function

overlap(f1::ScalarField,f2::ScalarField)

overlap(f1::VectorField,f2::VectorField)

overlap(m1::ScalarMode1D,m2::ScalarMode1D)

overlap(m1::ScalarMode2D,m2::ScalarMode2D)

overlap(f1::ScalarField,m2::ScalarMode2D)

overlap(m1::ScalarMode2D,f2::ScalarField)

overlap(m1::VectorMode,m2::VectorMode)

overlap(f1::VectorField,m2::VectorMode)

overlap(m1::VectorMode,f2::VectorField)

overlap(m1::ScalarModeFEM,m2::ScalarModeFEM)

overlap(m1::ScalarModeFEM,m2::ScalarfieldFEM)

overlap(m1::ScalarModeFEM,m2::ScalarfieldFEM)

overlap(m1::ScalarFieldFEM,m2::ScalarfieldFEM)

Effective area

The effective area is defined by:

for scalar modes: $\frac{\left( \iint \vert E \vert^2 dS \right)^2}{\iint \vert E \vert^4 dS}$

for vector modes, there are two different values [7] : $\frac{\mu_0}{\varepsilon_0}\frac{\left(Re\left(\iint(\vec{E}\wedge\vec{H}^*)_z dS\right)\right)^2}{\iint n_0^2 \vert \vec{E}.\vec{E}^* \vert^2 dS}$ and $\frac{\mu_0}{\varepsilon_0}\frac{\left(Re\left(\iint(\vec{E}\wedge\vec{H}^*)_z dS\right)\right)^2}{\iint n_0^2 \left(\vec{E}.\vec{E}\right)\left(\vec{E}^*.\vec{E}^*\right) dS}$ where $n_0$ is the refractive index involved in the relation between the nonlinear index $n_2$ and the third order susceptibility $\chi^{(3)}$.

OpticalFibers.ModeSolvers.Aeff — Function

Aeff(m::ScalarMode1D)

Aeff(m::ScalarMode2D)

Aeff(m::VectorMode,n0::Union{Real,Function}=0)

n0 must be a constant or a function of x and y.

If n0=0, this function assumes that n0≃neff (correct for weakly-guiding fibers).

Aeff(m::ScalarModeFEM)

Aeff(m::VectorModeFEM,n0::Union{Real,Function}=0)

n0 must be a constant or a function of x and y.

If n0=0, this function assumes that n0≃neff (correct for weakly-guiding fibers).

Non-linear coefficient

If all the lengths are in microns and $n_2$ is in SI (m²/W), you have to multiply the result by $10^{18}$ to obtain the result in W$^{-1}$.m$^{-1}$

OpticalFibers.ModeSolvers.nonLinearCoefficient — Function

nonLinearCoefficient(m::ScalarMode1D,n2::Union{Real,Function})

n2 must be a constant or a function of the radial coordinate r

nonLinearCoefficient(m::ScalarMode2D,n2::Union{Real,Function})

n2 must be a constant or a function of the cartesian coordinates x and y

nonLinearCoefficient(m::VectorMode,n2::Union{Real,Function},n0::Union{Real,Function}=0)

n2 and n0 must be a constant or a function of x and y.

If n0=0, this function assumes that n0≃neff (correct for weakly-guiding fibers).

nonLinearCoefficient(m::ScalarModeFEM,n2::Union{Real,Function})

n2 must be a constant or a function of the cartesian coordinates x and y

nonLinearCoefficient(m::VectorModeFEM,n2::Union{Real,Function},n0::Union{Real,Function}=0)

n2 and n0 must be constants or functions of x and y.

If n0=0, this function assumes that n0≃neff (correct for weakly-guiding fibers).

Mode Field Diameter (MFD)

The computation MFD is only valid for Gaussian-like beam (maximum at the center of the fiber and electric field with a constant sign). The MFD is calculated by finding the positions where $E=\frac{\max{\left(E\right)}}{\exp(1)}$ for scalar modes and $P_z=\frac{\max{\left(P_z\right)}}{\exp(2)}$ for vector modes. The optional parameters theta is the angle (in degrees) between the direction along which the MFD is computed and the x-axis.

OpticalFibers.ModeSolvers.MFD — Function

MFD(m::ScalarMode1D)

MFD(m::ScalarMode2D,theta::Real=0)

MFD(m::VectorMode,theta::Real=0)

MFD(m::ScalarModeFEM,theta::Real=0)

MFD(m::VectorModeFEM,theta::Real=0)