OpticalFibers.ModeSolvers - Modes and Fields

Fields

This module defines two types of fields: the scalar field (valid in the case of weakly guiding fibers) and the vector field. If the fiber has a cylindrical symmetry, it is sufficient to describe the variations of a scalar field along a radius (the 2D scalar field can be then obtained by multiplying the 1D field by $\cos(\ell \theta)$ or $\sin(\ell \theta)$ where $\ell$ is the azimuthal number).

Four abstract structures are defined to describe an electromagnetic field:

OpticalFibers.ModeSolvers.Field — Type

abstract type Field end

OpticalFibers.ModeSolvers.ScalarField1D — Type

abstract type ScalarField1D <: Field end

OpticalFibers.ModeSolvers.ScalarField2D — Type

abstract type ScalarField2D <: Field end

OpticalFibers.ModeSolvers.VectorField2D — Type

abstract type VectorField2D <: Field end

Each kind of field can be described by a function, a matrix (a vector in the 1D case) or a CellField when using a FEM solver.

Note that functions are always function of a tuple (r,) in the case of a 1D field or (x,y) in case of a 2D field. This ensures a direct compatibility with the packages Gridap(FEM method) and HCubature (for integration).

OpticalFibers.ModeSolvers.ScalarFieldFunction1D — Type

mutable struct ScalarFieldFunction1D <: ScalarField1D

nu :: Int - Azimuthal number

E :: Function

OpticalFibers.ModeSolvers.ScalarFieldFunction2D — Type

mutable struct ScalarFieldFunction2D <: ScalarField2D

E :: Function

OpticalFibers.ModeSolvers.VectorFieldFunction2D — Type

mutable struct VectorFieldFunction2D <: VectorField2D

Ex :: Function

Ey :: Function

Ez :: Function

Hx :: Function

Hy :: Function

Hz :: Function

OpticalFibers.ModeSolvers.ScalarFieldMatrix1D — Type

mutable struct ScalarFieldMatrix1D <: ScalarField1D

nu :: Int - Azymuthal number

r :: AbstractVector{<:Real}

E :: AbstractVector{<:Number}

OpticalFibers.ModeSolvers.ScalarFieldMatrix2D — Type

mutable struct ScalarFieldMatrix2D <: ScalarField2D

x :: AbstractVector{<:Real}

y :: AbstractVector{<:Real}

E :: AbstractMatrix{<:Number}

OpticalFibers.ModeSolvers.VectorFieldMatrix2D — Type

mutable struct VectorFieldMatrix2D <: VectorField2D

x :: AbstractVector{<:Real}

y :: AbstractVector{<:Real}

Ex :: AbstractMatrix{<:Number}

Ey :: AbstractMatrix{<:Number}

Ez :: AbstractMatrix{<:Number}

Hx :: AbstractMatrix{<:Number}

Hy :: AbstractMatrix{<:Number}

Hz :: AbstractMatrix{<:Number}

OpticalFibers.ModeSolvers.ScalarFieldFEM1D — Type

mutable struct ScalarFieldFEM1D <: ScalarField1D

nu :: Int - Azymuthal number

dΩ :: Gridap.CellData.Measure

E :: Gridap.CellField

OpticalFibers.ModeSolvers.ScalarFieldFEM2D — Type

mutable struct ScalarFieldFEM2D <: ScalarField2D

dΩ :: Gridap.CellData.Measure

E :: Gridap.CellField

OpticalFibers.ModeSolvers.VectorFieldFEM2D — Type

mutable struct VectorFieldMatrix2D <: VectorField2D

dΩ :: Gridap.CellData.Measure

Ex :: Gridap.CellField

Ey :: Gridap.CellField

Ez :: Gridap.CellField

Hx :: Gridap.CellField

Hy :: Gridap.CellField

Hz :: Gridap.CellField

The fields can be multiplied by a scalar and conjugated. It is also possible to take the real part or the imaginary part of a field. Two fields of the same type can also be added or subtracted.

A function is implemented to compute the value of a field at a given position (r,) in 1D or (x,y) in 2D:

OpticalFibers.ModeSolvers.computeField — Function

computeField(f::ScalarField1D,r::Union{Real,AbstractArray{<:Real}})

computeField(f::ScalarField2D,x::Union{Real,AbstractArray{<:Real}},y::Union{Real,AbstractArray{<:Real}})

computeField(f::VectorField2D,x::Union{Real,AbstractArray{<:Real}},y::Union{Real,AbstractArray{<:Real}},component::Symbol=:Ex)

computeField(m::Mode{<:ScalarField1D},r::Union{Real,AbstractArray{<:Real}},z::Real=0)

computeField(m::Mode{<:ScalarField2D},x::Union{Real,AbstractArray{<:Real}},y::Union{Real,AbstractArray{<:Real}},z::Real=0)

computeField(m::Mode{<:VectorField2D},x::Union{Real,AbstractArray{<:Real}},y::Union{Real,AbstractArray{<:Real}},component::Symbol=:Ex,z::Real=0)

Modes

A mode is a structure that contains a name, an effective index, the wavelength and, optionally, a field.

This package assumes that the field is proportionnal to $\exp\left(i\beta z-\omega t\right)$.

OpticalFibers.ModeSolvers.Mode — Type

struct Mode{T<:Union{Field,Nothing}}

Name :: String

neff :: Number

lambda :: Real

field :: Field or Nothing

Check if the field/mode is valid

A function was implemented to check if a field or a mode is valid by checking the number of arguments of the functions, the dimension of the matrices or the triangulaiton of the FEM fields.

OpticalFibers.ModeSolvers.isvalidField — Function

isvalidField(f::Field)

OpticalFibers.ModeSolvers.isvalidMode — Function

isvalidField(m::Mode)

Modes/Fields conversion

It is possible to convert a 1D function or matrix mode/field to 2d. The argument angle corresponds to the orientation in degrees (0 correspond to $\cos(\ell \theta)$ and -90 to $\sin(\ell \theta)$). This function is not available for FEM mode/field since it requires the creation of a 2D mesh.

OpticalFibers.ModeSolvers.convertTo2D — Function

convertTo2D(f::ScalarFieldFunction1D,angle::Real=0)

convertTo2D(m::Mode{ScalarFieldFunction1D},angle::Real=0)

convertTo2D(f::ScalarFieldMatrix1D,angle::Real=0)

convertTo2D(m::Mode{ScalarFieldMatrix1D},angle::Real=0)

It is also possible to convert a function or FEM field to a matrix mode/field.

OpticalFibers.ModeSolvers.convertToMatrix — Function

convertToMatrix(f::ScalarField1D,r::AbstractVector{<:Real})

convertToMatrix(m::Mode{<:ScalarField1D},r::AbstractVector{<:Real})

convertToMatrix(f::ScalarField2D,x::AbstractVector{<:Real},y::AbstractVector{<:Real})

convertToMatrix(m::Mode{<:ScalarField2D},x::AbstractVector{<:Real},y::AbstractVector{<:Real})

convertToMatrix(f::VectorField2D,x::AbstractVector{<:Real},y::AbstractVector{<:Real})

convertToMatrix(m::Mode{<:VectorField2D},x::AbstractVector{<:Real},y::AbstractVector{<:Real})

A mode that contains a ScalarField2D field can be converted to a mode with a VectorField2D with the approximation $\vec{k}=\beta \vec{e}_z$ and $E_z=H_z=0$.

OpticalFibers.ModeSolvers.convertToVector — Function

convertToVector(m::Mode{<:ScalarField2D},angle::Real=0)

Propagation of modes

The field due to the propagation of a mode after a propagation distance z can be calculated by using this function:

OpticalFibers.ModeSolvers.getField — Function

getField(m::Mode,z::Real=0)

Losses

The mode losses are computed from the imaginary part of the effective index.

The function returns the losses of a mode in dB/km if the wavelength is in meters. If the wavelength is in microns, you have to multiply the result by 1E6 to obtain the losses in dB/km.

OpticalFibers.ModeSolvers.losses — Method

losses(m::Mode)

Poynting vector

Convenient functions that compute the Poynting vector of 2D modes and fields. All functions return a tuple (Px,Py,Pz) that contains the three components of the Poynting vector in the same type as the field.

OpticalFibers.ModeSolvers.PoyntingVector — Function

PoyntingVector(f::VectorField2D)

PoyntingVector(m::Mode{<:ScalarField2D})

PoyntingVector(m::Mode{<:VectorField2D})

Mode normalization

The modes given by the mode solvers of this package are not normalized. You can normalize them by using the function normalize!. The vector modes are normalized so as $\vert \iint \frac{1}{2}\left(\vec{E}\wedge\vec{H}^*\right) dS \vert=1$ and the scalar mode are normalized like vector modes if the optional parameter unitIntegral is false or so as $ \iint \vert E \vert^2 dS=1 $ if unitIntegral is true.

OpticalFibers.ModeSolvers.normalize! — Function

normalize!(m::Mode{ScalarFieldFunction1D};unitIntegral::Bool=true,kwargs...)

normalize!(m::Mode{ScalarFieldFunction2D};unitIntegral::Bool=true,kwargs...)

normalize!(m::Mode{VectorFieldFunction2D};kwargs...)

normalize!(m::Mode{ScalarFieldMatrix1D};unitIntegral::Bool=true)

normalize!(m::Mode{ScalarFieldMatrix2D};unitIntegral::Bool=true)

normalize!(m::Mode{VectorFieldMatrix2D})

normalize!(m::Mode{ScalarFieldFEM1D};unitIntegral::Bool=true)

normalize!(m::Mode{ScalarFieldFEM2D};unitIntegral::Bool=true)

normalize!(m::Mode{VectorFieldFEM2D})

Overlap integral

The overlap integral between two fields or modes f$_1$ and f$_2$ is $\langle f_1 \vert f_2 \rangle = \iint E_1 E_2^* dS$ for scalar fields/modes and $\langle f_1 \vert f_2 \rangle = \iint \frac{1}{2}\left(\vec{E}_1 \wedge \vec{H}_2^*\right)_z dS$ for vector fields/modes.

OpticalFibers.ModeSolvers.overlap — Function

overlap(m1::ScalarFieldFunction1D,m2::ScalarFieldFunction1D;kwargs...)

overlap(f1::Union{ScalarField,ScalarMode},f2::Union{ScalarField,ScalarMode})

overlap(f1::Union{VectorField,VectorMode},f2::Union{VectorField,VectorMode})

overlap(m1::ScalarMode1D,m2::ScalarMode1D)

overlap(m1::ScalarModeFEM1D,m2::ScalarModeFEM1D)

overlap(m1::Union{ScalarFieldFEM,ScalarModeFEM},m2::Union{ScalarFieldFEM,ScalarModeFEM})

overlap(m1::Union{VectorFieldFEM,VectorModeFEM},m2::Union{VectorFieldFEM,VectorModeFEM})

Effective area

The effective area is defined by:

for scalar modes: $\frac{\left( \iint \vert E \vert^2 dS \right)^2}{\iint \vert E \vert^4 dS}$

for vector modes, there are two different values [10] : $\frac{\mu_0}{\varepsilon_0}\frac{\left(Re\left(\iint(\vec{E}\wedge\vec{H}^*)_z dS\right)\right)^2}{\iint n_0^2 \vert \vec{E}.\vec{E}^* \vert^2 dS}$ and $\frac{\mu_0}{\varepsilon_0}\frac{\left(Re\left(\iint(\vec{E}\wedge\vec{H}^*)_z dS\right)\right)^2}{\iint n_0^2 \left(\vec{E}.\vec{E}\right)\left(\vec{E}^*.\vec{E}^*\right) dS}$ where $n_0$ is the refractive index involved in the relation between the nonlinear index $n_2$ and the third order susceptibility $\chi^{(3)}$. If n0=0, the function assumes that n0≃neff (correct for weakly-guiding fibers).

OpticalFibers.ModeSolvers.Aeff — Function

Aeff(m::Mode{ScalarFieldFunction1D};kwargs...)

Aeff(m::Mode{ScalarFieldFunction2D};kwargs...)

Aeff(m::Mode{VectorFieldFunction2D},n0::Union{Real,Function}=0;kwargs...)

Aeff(m::Mode{ScalarFieldMatrix1D})

Aeff(m::Mode{ScalarFieldMatrix2D})

Aeff(m::Mode{VectorFieldMatrix2D},n0::Union{Real,Function}=0)

Aeff(m::Mode{ScalarFieldFEM1D})

Aeff(m::Mode{ScalarFieldFEM2D})

Aeff(m::Mode{VectorFieldFEM2D},n0::Union{Real,Function}=0)

Non-linear coefficient

If all the lengths are in microns and $n_2$ is in SI (m²/W), you have to multiply the result by $10^{18}$ to obtain the result in W$^{-1}$.m$^{-1}$. In this function, n2 is the nonlinear index and n0 is the refractive index involved in the relation between $n_2$ and the third order susceptibility $\chi^{(3)}$. They must be constants or functions of the tuple (r,) in the 1D case or (x,y) in the 2D case. If n0=0, the function assumes that n0≃neff (correct for weakly-guiding fibers). In the case of modes describes by functions, it is possible to specify the relative and the absolute tolerances for the integral calculations.

OpticalFibers.ModeSolvers.nonLinearCoefficient — Function

nonLinearCoefficient(m::Mode{ScalarFieldFunction1D},n2::Union{Real,Function};kwargs...)

nonLinearCoefficient(m::Mode{ScalarFieldFunction2D},n2::Union{Real,Function};kwargs...)

nonLinearCoefficient(m::Mode{VectorFieldFunction2D},n2::Union{Real,Function},n0::Union{Real,Function}=0;kwargs...)

nonLinearCoefficient(m::Mode{ScalarFieldMatrix1D},n2::Union{Real,Function})

nonLinearCoefficient(m::Mode{ScalarFieldMatrix2D},n2::Union{Real,Function})

nonLinearCoefficient(m::Mode{VectorFieldMatrix2D},n2::Union{Real,Function},n0::Union{Real,Function}=0)

nonLinearCoefficient(m::Mode{ScalarFieldFEM1D},n2::Union{Real,Function})

nonLinearCoefficient(m::Mode{ScalarFieldFEM2D},n2::Union{Real,Function})

nonLinearCoefficient(m::Mode{VectorFieldFEM2D},n2::Union{Real,Function},n0::Union{Real,Function}=0)

Mode Field Diameter (MFD)

The computation MFD is only valid for Gaussian-like beam (maximum at the center of the fiber and electric field with a constant sign). The MFD is calculated by finding the positions where $\vert E \vert=\frac{\max{\left(\vert E\vert \right)}}{\exp(1)}$ for scalar modes and $P_z=\frac{\max{\left(P_z\right)}}{\exp(2)}$ for vector modes. The optional parameters theta is the angle (in degrees) between the direction along which the MFD is computed and the x-axis. If the field is not quasi-Gaussian (in the case of a higher order mode for example), the value given by this function has no physical meaning.

OpticalFibers.ModeSolvers.MFD — Function

MFD(m::ScalarFieldFunction1D)

MFD(m::ScalarFieldFunction2D,theta::Real=0)

MFD(m::VectorFieldFunction2D,theta::Real=0)

MFD(m::ScalarFieldMatrix1D)

MFD(m::ScalarFieldMatrix2D,theta::Real=0)

MFD(m::VectorFieldMatrix2D,theta::Real=0)

MFD(m::ScalarFieldFEM1D)

MFD(m::ScalarFieldFEM2D,theta::Real=0)

MFD(m::VectorFieldFEM2D,theta::Real=0)

MFD(m::Mode{<:ScalarField1D})

MFD(m::Union{Mode{<:ScalarField2D},Mode{<:VectorField2D}},theta::Real=0)

Functions specific to FEM fields/modes

The function writevtk of the package Gridap is overloaded to save a FEM field or mode to a file that can be opened with ParaView.

Gridap.Visualization.writevtk — Function

Gridap.:writevtk(name::String,f::Union{ScalarFieldFEM2D,ScalarFieldFEM1D})

Gridap.:writevtk(name::String,f::VectorFieldFEM2D)

Gridap.:writevtk(name::String,m::Union{Mode{ScalarFieldFEM2D},Mode{ScalarFieldFEM1D},Mode{VectorFieldFEM2D}})

The function get_triangulation of the package Gridap is overloaded to get the triangulation of a FEM field (which is needed to plot a field with GridapMakie for example).

Gridap.Geometry.get_triangulation — Function

Gridap.:get_triangulation(m::Union{Mode{ScalarFieldFEM2D},Mode{ScalarFieldFEM1D},Mode{VectorFieldFEM2D}})

To compute the value of a CellField at a position (x,y), the function getValue is implemented to overcome the bug encountered in Gridap.jl when using an unregular triangular mesh (see issue 981). This function returns the value of the CellField at the positions given by the matrix of Gridap.Point p.

OpticalFibers.ModeSolvers.getValue — Function

getValue(f::Gridap.CellField,p::AbstractArray{<:Point})

getValue(f::Gridap.CellField,p::Point)