OpticalFibers.ModeSolvers - Solvers

Quasi-analytical multi-step index fibers solver

This solver assumes that the fiber consists of several concentric layers of uniform refractive index. The cladding is assumed to be infinite. In each layer, the analytical solutions are the Bessel functions. The solver uses the interface conditions to predict the effective index and the profile of the modes. This method is described in the book written by J. Bures [6].

This solver only returns guided modes: the effective index of the mode is real and cannot be lower than the refractive index of the external cladding.

OpticalFibers.ModeSolvers.multi_step_fiber_modes — Function

multi_step_fiber_modes(lambda::Real,nu::Integer,radius::Union{Vector{<:Real},Real},index::Vector{<:Real};maxPosition::Real=0,numberofPoints::Integer=100,precision::Float64=1E-12,type::Symbol=:Scalar,firstDivision::Integer=0)

Returns a vector of ScalarMode1D if type=:Scalar or a vector of VectorMode if type=:Vector. If maxPosition=0, the modes do not include the electromagnetic field.

lambda: wavelength

nu: azimuthal number

radius: outer radius of each layer (the cladding is inifinite and has no radius)

index: refractive index of each layer (the cladding is included so that length(index) must be equal to length(radius)+1)

maxPosition: maximum value of the radial coordinate r for scalar modes or cartesian coordinates x/y for vector modes. If maxPosition=0, the field is not computed.

numberofPoints: number of points between 0 and maxPosition

precision: precision required on the effective index

type: must be :Scalar or :Vector

firstDivision: in the case of a very multimode fiber, you can increase this number if some modes are missing. If firstDivision=0, the value of firstDivision is approximated by the solver.

Finite difference mode solvers

In this package, the FD solvers only return guided modes. The method used in the vectorial case is described in the paper of Zhu [7].

The computation of the effective index amounts to an eigenvalue problem. Three solutions are available to solve this eigenvalue problem: the use of the package Arpack.jl and the use of the package ArnoldiMethod.jl combined with LinearAlgebra.jl (LU decomposition) or MUMPS.jl.

OpticalFibers.ModeSolvers.FD — Function

FD(lambda::Real,l::Integer,mmax::Integer,fonc::Function,nb::Integer,rmax::Real;field::Bool=false,order::Integer=1,solver::Symbol=:Arpack,tol::Float64=0.0)

Returns a vector of ScalarMode1D in the case of a cylindrically-symmetric weakly-guiding fiber.

lambda: wavelength

l: azimuthal number

mmax: maximal number of modes (useful if the fiber is very multimode)

fonc: function of the radial coordinate r that describes the refractive index profile

nb: number of nodes for the finite difference method

rmax: maximal value of r

field: boolean that indicates if fields must be saved

order: order of the method to average the refractive index profile for each node (no averaging if order=1)

solver: can be :Arpack, :LU or :MUMPS

tol: tolerance for the eigenvalue solver (see documention of Arpack.jl or ArnoldiMethod.jl)

FD(lambda::Real,mmax::Integer,fonc::Function,nbx::Integer,nby::Integer,xmax::Real,ymax::Real;field::Bool=false,order::Integer=1,type::Symbol=:Scalar,solver::Symbol=:Arpack,tol::Float64=0.0)

Returns a vector of ScalarMode2D if type=:Scalar or a vector of VectorMode if type=:Vector.

lambda: wavelength

mmax: maximal number of modes (useful if the fiber is very multimode)

fonc: function of the cartesian coordinates x/y that describes the refractive index profile

nbx: number of nodes for the finite difference method in the direction x

nby: number of nodes for the finite difference method in the direction y

xmax: maximal value of x

ymax: maximal value of x

field: boolean that indicates if fields must be saved

order: order of the method to average the refractive index profile for each node (no averaging if order=1)

type: must be :Scalar or :Vector

solver: can be :Arpack, :LU or :MUMPS

tol: tolerance for the eigenvalue solver (see documention of Arpack.jl or ArnoldiMethod.jl)

Finite element mode solvers

The FEM solvers are based on Gridap.jl and can compute modes of isotropic and anisotropic fibers (useful when using a PML).

The computation of the effective index amounts to an eigenvalue problem. Two solutions are available to solve this eigenvalue problem: the use of the package ArnoldiMethod.jl combined with LinearAlgebra.jl (LU decomposition) or with MUMPS.jl.

In the case of anisotropic fibers, the dimension of the matrix is twice as large as in the case of isotropic fibers.

OpticalFibers.ModeSolvers.FEM — Function

FEM(lambda::Real,mmax::Int64,eps::Function,model::DiscreteModel,approx_neff::Real;order::Int64=2,field::Bool=false,type::Symbol=:Scalar,solver::Symbol=:LU,tol::Float64=0.0)

Returns a vector of ScalarModeFEM if type=:Scalar or a vector of VectorModeFEM if type=:Vector. The fiber is isotropic and described with its relative permittivity epsilon(pos) where pos is the tuple (x,y).

lambda: wavelength

mmax: number of modes to be calculated

eps: function of the tuple (x,y) that describes the relative permittivity profile

model: DiscreteModel generated with GridapGmsh.jl

approx_neff: effective index around which the modes will be computed

order: order of the FEM

field: boolean that indicates if fields must be saved

type: must be :Scalar or :Vector

solver: can be :LU or :MUMPS

tol: tolerance for the eigenvalue solver (see documention of ArnoldiMethod.jl)

FEM(lambda::Real,mmax::Int64,eps::tensor3,mu::tensor3,model::DiscreteModel,approx_neff::Real;order::Int64=1,field::Bool=false,solver::Symbol=:LU,tol::Float64=0.0)

Returns a vector of VectorModeFEM. The fiber is anisotropic and described with its relative permittivity tensor and its relative permeability tensor.

lambda: wavelength

mmax: number of modes to be calculated

eps: functions of the tuple (x,y) that describes the relative permittivity tensor profile

mu: functions of the tuple (x,y) that describes the relative permeability tensor profile

model: DiscreteModel generated with GridapGmsh.jl

approx_neff: effective index around which the modes will be computed

order: order of the FEM

field: boolean that indicates if fields must be saved

solver: can be :LU or :MUMPS

tol: tolerance for the eigenvalue solver (see documention of ArnoldiMethod.jl)