OpticalFibers.ModeSolvers - Solvers

Quasi-analytical multi-step index fibers solver

This solver assumes that the fiber consists of several concentric layers of uniform refractive index. The cladding is assumed to be infinite. In each layer, the analytical solutions are the Bessel functions. The solver uses the interface conditions to predict the effective index and the profile of the modes. This method is described in the book written by J. Bures [9].

This solver only returns guided modes: the effective index of the mode is real and cannot be lower than the refractive index of the outer cladding.

OpticalFibers.ModeSolvers.multi_step_fiber_modes — Function

multi_step_fiber_modes(lambda::Real,nu::Integer,radius::Union{Vector{<:Real},Real},index::Vector{<:Real};field::Bool=false,precision::Float64=1E-12,type::Symbol=:Scalar,firstDivision::Integer=0)

Returns a vector of Mode{ScalarFieldFunction1D} if type=:Scalar or a vector of Mode{VectorFieldFunction2D} if type=:Vector.

lambda: wavelength

nu: azimuthal number

radius: outer radius of each layer (the cladding is inifinite and has no radius)

index: refractive index of each layer (the cladding is included so that length(index) must be equal to length(radius)+1)

field: boolean that indicates if fields must be saved

precision: precision required on the effective index

type: must be :Scalar or :Vector

firstDivision: in the case of a very multimode fiber, you can increase this number if some modes are missing. If firstDivision=0, the value of firstDivision is approximated by the solver.

Finite element mode solvers

The FEM solvers are based on Gridap.jl and can compute modes of isotropic (functions FEM1D and FEM2D) and anisotropic fibers (function FEM2D_general). If a PML is not used, the functions FEM1D and FEM2D compute guided modes only. To compute the leaky modes, the user must add a PML by setting the value of dPML in these functions.

The computation of the effective index amounts to an eigenvalue problem which is solved by using the package ArnoldiMethod.jl combined with LinearAlgebra.jl (LU decomposition) or MUMPS.jl. Note that in the case of anisotropic fibers, the dimension of the matrix is twice as large as in the case of isotropic fibers.

OpticalFibers.ModeSolvers.FEM1D — Function

FEM1D(lambda::Real,nu::Int64,eps_fonc::Function,model::DiscreteModel;approx_neff::Real=0,neigs::Int64=1,order::Int64=2,field::Bool=false,solver::Symbol=:LU,tol::Real=0.0,verbose::Bool=false,dPML::Real=0,alphaPML::Real=10)

Returns a vector of Mode{ScalarFieldFEM1D}. The fiber is assumed to be isotropic with a cylindrical symmetry and is described with its relative permittivity.

lambda: wavelength

nu: azimuthal number

eps_fonc: function of r that describes the relative permittivity profile i.e. eps(r)=n(r)^2

model: DiscreteModel generated with GridapGmsh.jl

neigs: number of modes to compute

approx_neff: effective index around which the modes will be computed

order: order of the FEM

field: boolean that indicates if fields must be saved

solver: can be :LU or :MUMPS

tol: tolerance for the eigenvalue solver (see documention of ArnoldiMethod.jl)

verbose: boolean that enables some outputs

dPML: Thickness of the PML

alphaPML: attenuation coefficient of the PML

OpticalFibers.ModeSolvers.FEM2D — Function

FEM2D(lambda::Real,eps_fonc::Function,model::DiscreteModel;approx_neff::Real=0,neigs::Int64=1,order::Int64=2,field::Bool=false,solver::Symbol=:LU,tol::Float64=0.0,verbose::Bool=false,dPML::Real=0,alphaPML::Real=10,boundary_tag::String="",type::Symbol=:Scalar)

Returns a vector of Mode{ScalarFieldFEM2D} if type=:Scalar or a vector of Mode{VectorFieldFEM2D} if type=:Vector. The fiber is assumed to be isotropic and is described with its relative permittivity.

lambda: wavelength

eps_fonc: function of the tuple (x,y) that describes the relative permittivity profile i.e. eps(x)=n(x)^2

model: DiscreteModel generated with GridapGmsh.jl

neigs: number of modes to compute

approx_neff: effective index around which the modes will be computed

order: order of the FEM

field: boolean that indicates if fields must be saved

solver: can be :LU or :MUMPS

tol: tolerance for the eigenvalue solver (see documention of ArnoldiMethod.jl)

verbose: boolean that enables some outputs

dPML: Thickness of the PML

alphaPML: attenuation coefficient of the PML

boundary_tag: tag of the boundary in model. If "", then the function automatically detects the boundary.

type: :Scalar or :Vector

OpticalFibers.ModeSolvers.FEM2D_anisotropic — Function

FEM2D_anisotropic(lambda::Real,epsilon::tensor3,mu::tensor3,model::DiscreteModel;approx_neff::Real=0,neigs::Int64=1,order::Int64=1,field::Bool=false,solver::Symbol=:LU,tol::Float64=0.0,verbose::Bool=false,boundary_tag::String="")

Returns a vector of Mode{VectorFieldFEM2D}. The fiber is anisotropic and described with its relative permittivity tensor and its relative permeability tensor. The PML is assumed to be already included in the tensors epsilon and mu.

lambda: wavelength

epsilon: tensor3 with functions of (x,y) that describes the relative permittivity tensor profile

mu: tensor3 with functions of (x,y) that describes the relative permeability tensor profile

model: DiscreteModel generated with GridapGmsh.jl

approx_neff: effective index around which the modes will be computed

neigs: number of modes to calculate

order: order of the FEM

field: boolean that indicates if fields must be saved

solver: can be :LU or :MUMPS

tol: tolerance for the eigenvalue solver (see documention of ArnoldiMethod.jl)

verbose: boolean that enables some outputs

boundary_tag: tag of the boundary in model.

OpticalFibers.ModeSolvers.FEM2D_periodic — Function

FEM2D_periodic(lambda::Real,eps_fonc::Function,model::DiscreteModel;approx_neff::Real=0,neigs::Int64=1,order::Int64=2,field::Bool=false,solver::Symbol=:LU,tol::Real=0.0,verbose::Bool=false,kt::AbstractVector{<:Real}=[0.0,0.0],type::Symbol=:Scalar)

Returns a vector of Mode. The fiber is isotropic and described with its relative permittivity. The mesh must be periodic.

lambda: wavelength

epsilon: function of (x,y) that describes the relative permittivity profile

model: DiscreteModel generated with GridapGmsh.jl

approx_neff: effective index around which the modes will be computed

neigs: number of modes to calculate

order: order of the FEM

field: boolean that indicates if fields must be saved

solver: can be :LU or :MUMPS

tol: tolerance for the eigenvalue solver (see documention of ArnoldiMethod.jl)

verbose: boolean that enables some outputs

kt: vector of the 2D Brillouin zone

type: :Scalar or :Vector